backend/requirements/base.txt
django
djangorestframework
django-cors-headers
psycopg2-binary
python-dotenv
firebase-admin
django
djangorestframework
django-cors-headers
psycopg2-binary
python-dotenv
firebase-admin